requirements.txt
Django==3.1.0
django-heroku==0.3.1
gunicorn==20.1.0
numpy==1.19.4
tensorflow-cpu==2.5.0
Pillow==8.2.0